3.10 refactor: use indexes into the Sphere array list instead of pointers in the BVH implementation.
Storing indexes into the array reduces the memory footprint of the BVH.
We allocate a single array list for all of the sphere objects and simply
index into the array to retrieve the relevant information. The indices
array is allocated and populated once, when we construct our BVH. The
data defining the bvh node (i.e. the node's bounding box, as well as the
sphere index if the node is a leaf) are embedded into the Bvh node.
Links only store optional references to child links, which are used to
traverse the tree.

Starting at the root link, we recursively descend the bvh following the
child link pointers, checking if an incident ray intersects one of the
child bounding boxes. If the ray hits a bvh bounding box, then we check
the intersections within that bvh node. The bvh node (containing the
bounding box and optional sphere index data) is recovered from the
`*Link` using `@fieldParentPtr`. If the bvh node is a leaf we check if
the ray intersects the sphere the index stored in the node.

This implementation noticeably reduces the rendering time.