refactor: Use generic bounded array lists for BVH tree and sphere primitive list.
This change removes the HittableList container that previously stored
the scene primitive objects in an array list. Scene primitive objects
are now stored in the `BoundedList` generic, which is a thin wrapper
container around `std.ArrayList(T).initBuffer()` that implements a
bounded array list.

(See: https://ziglang.org/download/0.15.1/release-notes.html#Removal-of-BoundedArray)

Since all of scene objects are generated before rendering the image, we
can place an upper bound on the number of object primitives (and thus
the maximum number of BVH tree nodes) that must be allocated. The
bounded array list allows us pre-allocate the maximum capacity for the
underlying type, ensuring no reallocations are needed at runtime.